+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#678](https://github.com/alleslabs/celatone-frontend/pull/678) api v1 - contract table counts
+
 ### Improvements
 
 ### Bug fixes